Make User (de)serializable and restructure BookingSystem
The restructuring of `BookingSystem` was needed because each `User` has
a history of `Scorecard`s, which again has a `Course`. With the previous
design this would cause a circular reference, since each `Course` had a
`BookingSystem`. A `BookingSystem` now has a `Course` instead.

`Scorecard`s `HashMap` had to be changed as well, since it used to
contain a `User` - which introduced a circular reference in a `User`s
scorecard history.

`Object.equals` has also been overriden for several classes, to be ale
to check for equality after reading them from JSON.